+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2020-01-13
+
+- Update useDimensions to use `useLayoutEffect`
+- Remove `setIsModalActive` from gallery.tsx, derive it from `bookId` instead
+
 ## [1.0.6] - 2019-11-11
 
 - Update AboutModal image preload link